simplify habitCard
